+++ src/ui/buttons/button.jsx
@@ -109,7 +109,13 @@ export function Button({
   const colors = fundingConfig.colors;
   const secondaryColors = fundingConfig.secondaryColors || {};
 
-  let { color, period, label, shouldApplyRebrandedStyles } = style;
+  let {
+    color,
+    period,
+    label,
+    shouldApplyRebrandedStyles,
+    shouldApplyPayNowOrLaterLabel,
+  } = style;
 
   // if no color option is passed in via style props
   if (color === "" || typeof color === "undefined") {
@@ -189,6 +195,10 @@ export function Button({
           })
         : fundingConfig.labelText || fundingSource;
 
+    if (shouldApplyPayNowOrLaterLabel) {
+      labelText = "PayPal Pay Now or Later";
+    }
+
     if (!showPayLabel && instrument?.vendor && instrument.label) {
       labelText = instrument.secondaryInstruments
         ? `${instrument.secondaryInstruments[0].type} & ${instrument.vendor} ${instrument.label}`
@@ -234,6 +244,7 @@ export function Button({
       tagline={tagline}
       content={content}
       experiment={experiment}
+      shouldApplyPayNowOrLaterLabel={shouldApplyPayNowOrLaterLabel}
     />
   );
 
@@ -329,7 +340,9 @@ export function Button({
         tabindex="0"
         aria-label={labelText}
       >
-        <div class={CLASS.BUTTON_LABEL}>{labelNode}</div>
+        <div class={CLASS.BUTTON_LABEL} aria-hidden="true">
+          {labelNode}
+        </div>
 
         <Spinner />
       </div>