+++ test/integration/tests/index.js
@@ -3,6 +3,7 @@
 import "./checkout";
 import "./button";
 import "./wallet";
+import "./saved-payment-methods";
 import "./marks";
 import "./fields";
 import "./payment-fields";