+++ CHANGELOG.md
@@ -1,3 +1,99 @@
+## <small>5.0.413 (2026-05-04)</small>
+
+* feat: add variationName prop support for PayPal monogram marks (#2630) ([18a8e94](https://github.com/paypal/paypal-checkout-components/commit/18a8e94)), closes [#2630](https://github.com/paypal/paypal-checkout-components/issues/2630)
+
+
+
+## <small>5.0.412 (2026-04-27)</small>
+
+* feature(SavedPaymentMethods): Adding loading template for prerender (#2624) ([aba46b6](https://github.com/paypal/paypal-checkout-components/commit/aba46b6)), closes [#2624](https://github.com/paypal/paypal-checkout-components/issues/2624) [#2605](https://github.com/paypal/paypal-checkout-components/issues/2605)
+
+
+
+## <small>5.0.411 (2026-04-08)</small>
+
+* chore(release): 5.0.410 🎉 ([6d3ba5e](https://github.com/paypal/paypal-checkout-components/commit/6d3ba5e))
+
+
+
+## <small>5.0.410 (2026-04-08)</small>
+
+* feature(SavedPaymentMethods): New component SavedPaymentMethods (#2604) ([2eed0ef](https://github.com/paypal/paypal-checkout-components/commit/2eed0ef)), closes [#2604](https://github.com/paypal/paypal-checkout-components/issues/2604)
+
+
+
+## <small>5.0.409 (2026-04-03)</small>
+
+* Revert "Pass bfcache events via post-robot (#2608)" (#2616) ([bc56fea](https://github.com/paypal/paypal-checkout-components/commit/bc56fea)), closes [#2608](https://github.com/paypal/paypal-checkout-components/issues/2608) [#2616](https://github.com/paypal/paypal-checkout-components/issues/2616)
+
+
+
+## <small>5.0.408 (2026-03-23)</small>
+
+* Pass bfcache events via post-robot (#2608) ([2e278df](https://github.com/paypal/paypal-checkout-components/commit/2e278df)), closes [#2608](https://github.com/paypal/paypal-checkout-components/issues/2608)
+
+
+
+## <small>5.0.407 (2026-03-18)</small>
+
+* [DTINAPPXO-3543] Pass partnerAttributionID (#2607) ([8c201d8](https://github.com/paypal/paypal-checkout-components/commit/8c201d8)), closes [#2607](https://github.com/paypal/paypal-checkout-components/issues/2607)
+
+
+
+## <small>5.0.406 (2026-03-03)</small>
+
+* add label "Pay now or later" experiment (#2600) ([6c0d3e3](https://github.com/paypal/paypal-checkout-components/commit/6c0d3e3)), closes [#2600](https://github.com/paypal/paypal-checkout-components/issues/2600)
+
+
+
+## <small>5.0.405 (2026-02-26)</small>
+
+* Additional fix for hash params in resume flow (#2602) ([0c53815](https://github.com/paypal/paypal-checkout-components/commit/0c53815)), closes [#2602](https://github.com/paypal/paypal-checkout-components/issues/2602)
+
+
+
+## <small>5.0.404 (2026-02-17)</small>
+
+* chore: prefer org-level code of conduct file (#2595) ([e4433f6](https://github.com/paypal/paypal-checkout-components/commit/e4433f6)), closes [#2595](https://github.com/paypal/paypal-checkout-components/issues/2595)
+
+
+
+## <small>5.0.403 (2026-02-03)</small>
+
+* Revert "Add support for disabling buttons (#2590)" (#2593) ([1e48e9f](https://github.com/paypal/paypal-checkout-components/commit/1e48e9f)), closes [#2590](https://github.com/paypal/paypal-checkout-components/issues/2590) [#2593](https://github.com/paypal/paypal-checkout-components/issues/2593)
+
+
+
+## <small>5.0.402 (2026-01-27)</small>
+
+* Add support for disabling buttons (#2590) ([14ca952](https://github.com/paypal/paypal-checkout-components/commit/14ca952)), closes [#2590](https://github.com/paypal/paypal-checkout-components/issues/2590)
+
+
+
+## <small>5.0.401 (2026-01-22)</small>
+
+* App Switch in New Tab (#2581) ([fcc4629](https://github.com/paypal/paypal-checkout-components/commit/fcc4629)), closes [#2581](https://github.com/paypal/paypal-checkout-components/issues/2581)
+
+
+
+## <small>5.0.400 (2026-01-15)</small>
+
+* Revert "feat(buttons): add disabled prop support (#2584)" (#2588) ([a31cab9](https://github.com/paypal/paypal-checkout-components/commit/a31cab9)), closes [#2584](https://github.com/paypal/paypal-checkout-components/issues/2584) [#2588](https://github.com/paypal/paypal-checkout-components/issues/2588)
+
+
+
+## <small>5.0.399 (2026-01-06)</small>
+
+* fix: use the right env variable name for npm (#2586) ([e272f5b](https://github.com/paypal/paypal-checkout-components/commit/e272f5b)), closes [#2586](https://github.com/paypal/paypal-checkout-components/issues/2586)
+
+
+
+## <small>5.0.398 (2026-01-06)</small>
+
+* feat(buttons): add disabled prop support (#2584) ([5b9c8cd](https://github.com/paypal/paypal-checkout-components/commit/5b9c8cd)), closes [#2584](https://github.com/paypal/paypal-checkout-components/issues/2584)
+
+
+
 ## <small>5.0.397 (2025-11-06)</small>
 
 * remove the restriction to use PayPal urls (#2578) ([3444b38](https://github.com/paypal/paypal-checkout-components/commit/3444b38)), closes [#2578](https://github.com/paypal/paypal-checkout-components/issues/2578)